feat: Start rewriting Flyer.pde class in p5.js
This is the initial commit towards reproducing the Processing
water_spout scketch entirely in javascript with p5.js. This
commit, reproduces most of the original water_spout's Flyer class.
Changes include extensive use of p5.Vector class. Also I will try
not to hard code any object dimensions, will try to base them in
terms of width and height of the canvas. In this case, for the flyer,
we have its diameter D defined as 0.035 of the canvas diagonal.
TODO: Add flyer tracking to pressed mouse