feat: Implement drag from rim and begin adding drops to orb
Vectors to the rescue once again, now with the dragging operation.
Added drag_start parameter to Orb to record the position on the
rim the orb was hooked. To drag simply define a translation by the
difference of that start point and the mouse, then use it to move
the orb center. I still have not decided which parameters will be
used to create the orb itself, as of now I have changed intialization
parameters to just be the number of drops, color should probably
be another. TODO: Make drops pulsate as in original sketch.