feat: Write express server to handle media loading
Running the app directly on the browser caused it to
crash when trying to load sound files. Apparently in
order to have more complex behavior like this it must
be run on a sever, as "cross-origin" errors are reported
by the browser. I restructured the file lay-out to
correspond to that of a proper app intended to be hosted
on a server. Having already forgotten most of what I at
some point learned looking into MEAN development, I am
following p5JS' own, Daniel Shiffman, tutorial on basic
setup:

https://www.youtube.com/watch?v=2hhEOGXcCvg

Files now load and I tested a sound playback. TODO: At
this point might as well go all the why and write collaborative
functionality as is being explained in Shiffman's tutorial of
awesomeness.