feat: Implemented white drop flash that keeps beat
The idea here was the same that I had on the original sketch
where their is a continous (floating point) increment of the
beat index of 1/beat_time. You then floor it to use as an actual
index in the drop position array. In this way over the duration of
the beat time the drop will remain white. Or so I was reasoning.
However I get significantly different behavior if I update the beat
index before, as opposed to after, animating the whole drop array.
Where the pre-update is closer to the description above, and the
post-update speeds up the illusion of movement. I think it has to
do with the fact that you only "see" the last portion of the time
interval. I have kept the post-animation behavior as it is closer
to what "feels" better.