fix: Have each drop own its own sound file instance
Since the orb was using one instance of a loaded sound file it was
difficult to coordinate cueing play in accordance to the active drops,
since inevitably the sound file would be playing at the same time other
drops were cueing play. Rather than devising fine time control, the
array drop_active is replaced with drops, which holds both the status
of each drop (whether it is active i.e. clicked on) and a loaded instance
of the sound file it should play. In this way I can take advantage of
p5.sound.js' built in isPlaying function to que play only once when
the beat reaches the active drop, and additionally different active
drops play independently.