fix: Remove drop_in_time increment from drop rendering for loop
I found the reason for the difference in behavior between what I
was calling "pre" and "post" drop_in_time updates. I had not noticed
that I was incrementing inside the drop rendering for loop being
accustomed to python indentation rules I was fooled indentation
thinking it was scope. Since the increment was happening for every
drop, the beat was much faster. This rate can be achieved by simply
reducing beat_time. So beats can be reasoned in terms of sub divisions
of the frame rate as was originally intended. For example at 60 fps
with beat_time = 6 frames, draw is called 60 times in 1 second, since
the orb beats every 6 frames, there should be 10 orb beats in 1 second.
It seems to be more or less true by inspection.