feat: Display columns of Orb drum options
The goal is to have each sound file be represented by an
Orb, and drag Orbs from the central screen and swap them
with the ones on the columns on the left, whenever you
seek to integrate a new drum. I am not sure exactly how
to accomplish this functionality. Its turning out to be
trickier than I thought it would be.