refactor: Give an ordering to sketch orb objects
So as to organize the sketch in a saner manner, the order imposed
by the drum sound files is now passed on to all the orb objects in
the sketch. I added a orb_centers array holding the center of each
orb in the screen layout as well. The idea is that since there is
now an index for each orb, I can have a swapOrb function that can
swap orbs realtive to their index. In this way the user will be able
to switch between the main sector orbs and those on the columns
to the right.